Improved image animation to prevent flickering with slow internet connection
