+++ src/css/homepage.css
@@ -5,7 +5,6 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    overflow: hidden;
     z-index: 1;
     background-color: black;
 }
@@ -18,20 +17,10 @@
     height: 100%;
     object-fit: cover;
     opacity: 0;
-    transition: opacity 1s ease-in-out;
+    transition: opacity 3s ease-in-out;
     z-index: -1;
     pointer-events: none;
     filter: blur(0.5vh);
-}
-
-.fade-in {
-    opacity: 1;
-    z-index: -2;
-}
-
-.fade-out {
-    opacity: 0;
-    z-index: -1;
 }
 
 .visible {
@@ -39,7 +28,7 @@
 }
 
 .hidden {
-    display: none;
+    opacity: 0;
 }
 
 .card {
